requirements.txt
pytesseract
pillow
pandas
opencv-python
numpy
